Read C and C++ with the grammar, the scanner, the generator, and the runtime of the fork
This commit holds each change of the fork on top of the upstream tree-sitter-cpp
that the parent commit names. The work is one commit, because each earlier commit
carried a copy of src/parser.c that reached 468 MB, and a repository with such a
file cannot go to GitHub.

THE GRAMMAR

grammar/src holds the grammar of C and C++ in a Rust DSL, in place of grammar.js.
`cargo xtask generate` writes src/grammar.json, src/parser.c, src/node-types.json,
and src/tree_sitter. xtask also runs the corpus tests, the syntax snippets, a fuzz
check of the scanner, a parse of a corpus, and a comparison with the Clang AST.

The grammar reads the syntax of C++17 through C++26 and the papers of C++29 that
have wording: the reflection operator and splices, pack indexing, contracts,
expansion statements, `if consteval`, explicit object parameters, and the
requires expression in each of its positions. It also reads the extensions of
GCC, of Clang, and of MSVC, and the macro forms that appear in real code, for
example an attribute macro before a declaration, a macro that gives a type, and a
line that holds only macro invocations.

THE SCANNER

src/scanner.c reads what a token pattern cannot: the preprocessor lines and their
conditional groups, the text of a directive line with its raw strings and its line
splices, the macro invocation forms, the line endings of DOS and of the old Mac,
and the record of the type names that a file declares.

THE MEASUREMENT

A corpus of 329,387 files from 64 codebases gives 14,256 files with a parse error.
The compiler test suites of GCC 16.2 and of Clang 22.1 give 7,646 files, and the
trees agree with the Clang AST on 98.8 percent of 206,072 facts.

THE RUNTIME AND THE GENERATOR

vendor/tree-sitter holds the runtime at ABI 1016, and vendor/tree-sitter-generate
holds the generator. The runtime keeps each repair of the fork and it still reads
each upstream grammar of ABI 14 and ABI 15, so a consumer can load this grammar
next to the grammars of crates.io.

upstream-patches holds one patch for each repair that upstream can take, against
the tag v0.27.0. They give error recovery a stack summary of depth 256, a cost for
an unmatched bracket, a recovery to three previous states, and a merge of the
versions that the GLR version limit would otherwise remove. That last one repairs
a silent misparse: the limit removed the version that carried the reading which
the dynamic precedence prefers, so a deeply nested template argument read as an
expression, and a whole declaration read as an expression statement, each with no
ERROR node.

THE PARSE TABLES

The tables use 32-bit values, because the grammar passed the 16-bit limit of
65,535 actions. They also use a shape layout: the states that share one set of
symbol groups hold their symbols one time, and a column whose value is the same
for each such state holds that value one time. src/parser.c falls from 468,520,078
bytes to 11,609,554, a factor of 40. The read-only data of the object file falls
from 232 MB to 7.5 MB, and the compile of that file falls from 24.5 s and 2.99 GB
to 2.7 s and 0.38 GB.

WHAT IS NOT HERE

The parser has no name lookup, so it cannot tell a type name from a value name.
`T(x)` and `f(x)` have one token shape, and the tree reads a call. That class holds
the largest part of the remaining disagreement with the Clang AST.

The source of this commit is the vendor part of tree-sitter-cpp commit 5882494fb792f030f048c75bdc0b452f42a6a1c1.

+++ crates/generate/src/build_tables.rs
@@ -123,7 +123,9 @@ pub fn build_tables(
         );
     }
 
-    if parse_table.states.len() > u16::MAX as usize {
+    // A state id has 32 bits, and the runtime uses UINT32_MAX for a subtree with no state
+    // (tree-sitter-cpp fork).
+    if parse_table.states.len() > u32::MAX as usize {
         Err(ParseTableBuilderError::StateCount(parse_table.states.len()))?;
     }
